requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
backcall==0.1.0           # via ipython
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
decorator==4.3.2          # via ipython, traitlets
django-rest-auth==0.9.3
django==2.1.7
djangorestframework==3.9.1
entrypoints==0.3          # via flake8
flake8==3.7.7
idna==2.8                 # via requests
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.3.0            # via ipdb
jedi==0.13.3              # via ipython
mccabe==0.6.1             # via flake8
more-itertools==6.0.0     # via pytest
parso==0.3.4              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.9.0             # via pytest
prompt-toolkit==2.0.9     # via ipython
psycopg2==2.7.7
ptyprocess==0.6.0         # via pexpect
py==1.8.0                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.0           # via flake8
pygments==2.3.1           # via ipython
pytest==4.3.0
pytz==2018.9              # via django
requests==2.21.0
six==1.12.0               # via django-rest-auth, prompt-toolkit, pytest, traitlets
traitlets==4.3.2          # via ipython
urllib3==1.24.1           # via requests
wcwidth==0.1.7            # via prompt-toolkit
